Move redirect handling
