+++ services/graphql-server/src/graphql/definitions/website/index.js
@@ -1,6 +1,7 @@
 const gql = require('graphql-tag');
 const inquirySubmission = require('./inquiry-submission');
 const option = require('./option');
+const redirect = require('./redirect');
 const schedule = require('./schedule');
 const section = require('./section');
 const site = require('./site');
@@ -9,6 +10,7 @@ module.exports = gql`
 
 ${inquirySubmission}
 ${option}
+${redirect}
 ${schedule}
 ${section}
 ${site}
