+++ services/graphql-server/src/graphql/definitions/website/site.js
@@ -7,7 +7,6 @@ extend type Query {
   websiteSite(input: WebsiteSiteQueryInput!): WebsiteSite @findOne(model: "platform.Product", using: { id: "_id" }, criteria: "websiteSite")
   websiteSites(input: WebsiteSitesQueryInput = {}): WebsiteSiteConnection! @findMany(model: "platform.Product", criteria: "websiteSite")
   matchWebsiteSites(input: MatchWebsiteSitesQueryInput!): WebsiteSiteConnection! @matchMany(model: "platform.Product", criteria: "websiteSite")
-  websiteRedirect(input: WebsiteRedirectQueryInput!): WebsiteRedirect
 }
 
 type WebsiteSite {
@@ -54,12 +53,6 @@ enum WebsiteSiteMatchField {
   name
 }
 
-type WebsiteRedirect {
-  from: String!
-  to: String!
-  code: Int!
-}
-
 type WebsiteSiteConnection @projectUsing(type: "WebsiteSite") {
   totalCount: Int!
   edges: [WebsiteSiteEdge]!
@@ -87,12 +80,6 @@ enum WebsiteSiteSortField {
   id
   name
   fullName
-}
-
-input WebsiteRedirectQueryInput {
-  siteId: ObjectID
-  from: String!
-  params: JSON
 }
 
 input WebsiteSiteQueryInput {
